Update import statements to match renamed files
