+++ src/modules/markets/container.js
@@ -10,7 +10,7 @@ import { updateMarketsFilteredSorted, clearMarketsFilteredSorted } from 'modules
 import { toggleFavorite } from 'modules/markets/actions/update-favorites'
 
 import loadMarkets from 'modules/markets/actions/load-markets'
-import { loadMarketsByTopic } from 'modules/markets/actions/load-markets-by-topic'
+import { loadMarketsByTopic } from 'modules/markets/actions/load-markets-by-category'
 import { loadMarketsInfo } from 'modules/markets/actions/load-markets-info'
 
 import getValue from 'utils/get-value'
