+++ test/categories/actions/update-has-loaded-category-test.js
@@ -1,7 +1,7 @@
 import { describe, it } from 'mocha'
 import { assert } from 'chai'
 
-import { updateHasLoadedTopic, UPDATE_HAS_LOADED_TOPIC } from 'modules/topics/actions/update-has-loaded-topic'
+import { updateHasLoadedTopic, UPDATE_HAS_LOADED_TOPIC } from 'modules/categories/actions/update-has-loaded-category'
 
 describe('modules/topics/actions/update-has-loaded-topic.js', () => {
   const test = (t) => {
